% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edgebundleR-package.R
\name{edgeToJSON}
\alias{edgeToJSON}
\title{Helper function to convert edges to JSON}
\usage{
edgeToJSON(edges)
}
\arguments{
\item{edges}{a matrix of edge relationships}
}
\description{
Helper function to convert edges to JSON
}

