% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edgebundleR-package.R
\name{edgeToJSON_igraph}
\alias{edgeToJSON_igraph}
\title{Helper function to convert an igraph to JSON}
\usage{
edgeToJSON_igraph(graph)
}
\arguments{
\item{graph}{an igraph}
}
\description{
Helper function to convert an igraph to JSON
}

