% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgebundleR-package.R
\name{edgeToJSON_matrix}
\alias{edgeToJSON_matrix}
\title{Helper function to convert edges to JSON}
\usage{
edgeToJSON_matrix(edges)
}
\arguments{
\item{edges}{a matrix of edge relationships}
}
\description{
Helper function to convert edges to JSON
}

