% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/edgebundleR.R
\name{renderEdgebundle}
\alias{renderEdgebundle}
\title{Widget render function for use in Shiny}
\usage{
renderEdgebundle(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

