% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/saveEdgebundle.R
\name{saveEdgebundle}
\alias{saveEdgebundle}
\title{Save a edge bundle to an HTML file}
\usage{
saveEdgebundle(x, file, selfcontained = TRUE)
}
\arguments{
\item{x}{plot to save (e.g. result of calling the function
  \code{edgebundle}).}

\item{file}{File to save HTML into}

\item{selfcontained}{Whether to save the HTML as a single self-contained file
(with external resources base64 encoded) or a file with external resources
placed in an adjacent directory.}
}
\description{
Save a edge buyndle graph to an HTML file for sharing with others. The HTML can
include it's dependencies in an adjacent directory or can bundle all
dependencies into the HTML file (via base64 encoding).
}

