% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyedge.R
\name{shinyedge}
\alias{shinyedge}
\title{Shiny interface to the edgebundle function}
\usage{
shinyedge(x)
}
\arguments{
\item{x}{an appropriately structured JSON file (see vignette for details) or a
  square symmetric matrix (e.g. correlation matrix) or an igraph object.}
}
\description{
Opens a shiny GUI to facilitate interaction with the edgebundle function
}
\details{
Circle plot with bundled edges

Takes an appropriately structured JSON file or a square symmetric matrix (e.g. a
correlation matrix or precision matrix) and outputs a circle plot with the nodes
around the circumfrence and linkages between the connected nodes. Adapted from
the  Mike Bostock's D3 Hierarchical Edge Bundling example using the htmlwidgets
framework.
}

