% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edgebundleR-package.R
\name{adjToEdge}
\alias{adjToEdge}
\title{Helper function to convert adjacency matrix to edges}
\usage{
adjToEdge(adj)
}
\arguments{
\item{adj}{an adjacency matrix}
}
\description{
Helper function to convert adjacency matrix to edges
}

