% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edgebundleR.R
\name{edgebundle}
\alias{edgebundle}
\title{Circle plot with bundled edges}
\usage{
edgebundle(x, tension = 0.5, cutoff = 0.1, width = NULL, fontsize = 14,
  padding = 100, maxsize = 10)
}
\arguments{
\item{x}{an appropriately structured JSON file (see vignette for details) or a
square symmetric matrix (e.g. correlation matrix) or an igraph object.}

\item{tension}{numeric between 0 and 1 giving the tension of the links}

\item{cutoff}{numeric giving the threshold dependence for linkages to be plotted}

\item{width}{the width of the plot when viewed externally}

\item{fontsize}{font size of the node labels}

\item{padding}{the padding (in px) between the inner radius of links and the
  edge of the plot.  Increase this when the labels run outside the edges of
  the plot.  Default: 100.}
}
\description{
Takes an appropriately structured JSON file or a square symmetric matrix (e.g. a
correlation matrix or precision matrix) and outputs a circle plot with the nodes
around the circumfrence and linkages between the connected nodes. Adapted from
the  Mike Bostock's D3 Hierarchical Edge Bundling example using the htmlwidgets
framework.
}
\examples{
\dontrun{
require(igraph)
ws_graph = watts.strogatz.game(1, 50, 4, 0.05)
edgebundle(ws_graph,tension = 0.1,fontsize = 20)
}
}

