% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edgebundleR.R
\name{renderEdgebundle}
\alias{renderEdgebundle}
\title{Widget render function for use in Shiny}
\usage{
renderEdgebundle(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{edgebundle expression}

\item{env}{environment}

\item{quoted}{logical, default = FALSE}
}
\description{
Widget render function for use in Shiny
}

