% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyedge.R
\name{shinyedge}
\alias{shinyedge}
\title{Shiny interface to the edgebundle function}
\usage{
shinyedge(x)
}
\arguments{
\item{x}{an appropriately structured JSON file (see vignette for details) or a
square symmetric matrix (e.g. correlation matrix) or an igraph object.}
}
\description{
Opens a shiny GUI to facilitate interaction with the edgebundle function
}

